requirements.txt
clip
imageio
cython
imageio-ffmpeg
kornia
numpy
Pillow
PyGLM
resize-right
scipy
smplx
tqdm
Ninja
pyyaml
matplotlib
trimesh
tensorboard
easydict  
fire
pymeshlab
libigl==2.4.1
cholespy
torch-sparse
torch-scatter
